make ValidatorIndex a u64

+++ src/altair/state_transition/block_processing.rs
@@ -86,7 +86,7 @@ pub fn process_attester_slashing<
     let mut slashed_any = false;
     let current_epoch = get_current_epoch(state, context);
     for &index in &indices {
-        if is_slashable_validator(&state.validators[index], current_epoch) {
+        if is_slashable_validator(&state.validators[index as usize], current_epoch) {
             slash_validator(state, index, None, context)?;
             slashed_any = true;
         }
@@ -176,7 +176,7 @@ pub fn process_block_header<
         body_root: block.body.hash_tree_root()?,
         ..Default::default()
     };
-    let proposer = &state.validators[block.proposer_index];
+    let proposer = &state.validators[block.proposer_index as usize];
     if proposer.slashed {
         return Err(invalid_header_error(
             InvalidBeaconBlockHeader::ProposerSlashed(proposer_index),
@@ -339,11 +339,14 @@ pub fn process_proposer_slashing<
         )));
     }
     let proposer_index = header_1.proposer_index;
-    let proposer = state.validators.get(proposer_index).ok_or_else(|| {
-        invalid_operation_error(InvalidOperation::ProposerSlashing(
-            InvalidProposerSlashing::InvalidIndex(proposer_index),
-        ))
-    })?;
+    let proposer = state
+        .validators
+        .get(proposer_index as usize)
+        .ok_or_else(|| {
+            invalid_operation_error(InvalidOperation::ProposerSlashing(
+                InvalidProposerSlashing::InvalidIndex(proposer_index),
+            ))
+        })?;
     if !is_slashable_validator(proposer, get_current_epoch(state, context)) {
         return Err(invalid_operation_error(InvalidOperation::ProposerSlashing(
             InvalidProposerSlashing::ProposerIsNotSlashable(header_1.proposer_index),
@@ -403,7 +406,7 @@ pub fn process_randao<
 ) -> Result<()> {
     let mut epoch = get_current_epoch(state, context);
     let proposer_index = get_beacon_proposer_index(state, context)?;
-    let proposer = &state.validators[proposer_index];
+    let proposer = &state.validators[proposer_index as usize];
     let domain = get_domain(state, DomainType::Randao, Some(epoch), context)?;
     let signing_root = compute_signing_root(&mut epoch, domain)?;
     if verify_signature(
@@ -451,7 +454,7 @@ pub fn process_voluntary_exit<
     let voluntary_exit = &mut signed_voluntary_exit.message;
     let validator = state
         .validators
-        .get(voluntary_exit.validator_index)
+        .get(voluntary_exit.validator_index as usize)
         .ok_or_else(|| {
             invalid_operation_error(InvalidOperation::VoluntaryExit(
                 InvalidVoluntaryExit::InvalidIndex(voluntary_exit.validator_index),
